src/test/java/jpabook/jpashop/Repository/MemberRepositoryTest.java
package jpabook.jpashop.Repository;

import jpabook.jpashop.domain.Member;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.Assert.*;

@SpringBootTest
public class MemberRepositoryTest {
    @Autowired
    MemberRepository memberRepository;
    @Test@Transactional
    public void testMember(){
        Member member =new Member();
        member.setName("TEST123");
        Long saveId=memberRepository.save(member);
        Member findMember=memberRepository.findMember(saveId);
        Assertions.assertThat(findMember.getId()).isEqualTo(member.getId());
        Assertions.assertThat(findMember.getName()).isEqualTo(member.getName());

    }


}